frmMain.py -- small fix for bug introduced with autosave feature

+++ src/ui/frmMain.py
@@ -2292,7 +2292,8 @@ class frmMain(QMainWindow, Ui_frmMain):
 
         if self.project_settings.general_preferences['AutoBackup'] > 0:
             from shutil import copyfile
-            copyfile(file_name, file_name + ".bak")
+            if os.path.exists(file_name):
+                copyfile(file_name, file_name + ".bak")
 
         project_writer = self.project_writer_type()
         project_writer.write_file(self.project, file_name)
@@ -2342,6 +2343,7 @@ class frmMain(QMainWindow, Ui_frmMain):
             path_only, file_only = os.path.split(file_name)
             try:
                 self.save_project(file_name)
+                self.project.file_name = file_name
                 self.setWindowTitle(self.model + " - " + file_only)
                 if path_only != directory:
                     self.program_settings.setValue("ProjectDir", path_only)
